msg: Make get_file_line return a size_t
The underlying API call to po_filepos_start_line returns a size_t.

The intent of returning a pointer was to return `NULL` in case of errors
but the current code was not doing any such NULL-check. We take
advantage of the fact that 0 is not a valid line number to simplify this
and add those checks.

+++ src/gtr-context.h
@@ -36,6 +36,8 @@ G_DECLARE_FINAL_TYPE (GtrContextPanel, gtr_context_panel, GTR, CONTEXT_PANEL, Ad
  */
 GtkWidget        *gtr_context_panel_new                   (void);
 void              gtr_context_init_tm                     (GtrContextPanel *panel, GtrTranslationMemory *tm);
-void              gtr_context_add_path                    (GtrContextPanel *panel, const char *filename, int line);
+void              gtr_context_add_path                    (GtrContextPanel *panel,
+                                                           const char      *filename,
+                                                           size_t           line);
 
 G_END_DECLS
